Use the get_wch because it's possible to get special chars
I'm a fucking brazillian, I want to type ç and é in the terminal.

May the action input will need to have a representation to
distinguish special keys like arrow keys and special chars, because
the Ă and the arrow key down has the same char code.

And the get_wch return an string given a typed char and a integer
when it is a special key. I think that the type of variable is not
the better way to do that.