Introduce a new render system that run in a poll of one thread.
The idea is that the render always should be executed but it just
schedule the render task and set the screen to be rendered.

If task always will render the last screen. It will prevent the render
to buffer the screens to be rendered

This commit add verbosity to the travis. Now it will show the tests
cases.